Move support out of tpbot package to highlight it is a support and not domain specific.
