+++ TelepresenceBot/core/src/main/java/com/novoda/support/Observer.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.support;
+package com.novoda.support;
 
 public interface Observer<T> {
