+++ TelepresenceBot/core/src/main/java/com/novoda/tpbot/bot/BotTelepresenceService.java
@@ -2,7 +2,7 @@ package com.novoda.tpbot.bot;
 
 import com.novoda.tpbot.Direction;
 import com.novoda.tpbot.Result;
-import com.novoda.tpbot.support.Observable;
+import com.novoda.support.Observable;
 
 interface BotTelepresenceService {
