+++ TelepresenceBot/core/src/main/java/com/novoda/tpbot/human/HumanPresenter.java
@@ -2,10 +2,10 @@ package com.novoda.tpbot.human;
 
 import com.novoda.tpbot.Direction;
 import com.novoda.tpbot.Result;
-import com.novoda.tpbot.support.Observable;
-import com.novoda.tpbot.support.Observer;
+import com.novoda.support.Observable;
+import com.novoda.support.Observer;
 
-import static com.novoda.tpbot.support.Observable.unsubscribe;
+import static com.novoda.support.Observable.unsubscribe;
 
 class HumanPresenter {
