+++ TelepresenceBot/core/src/main/java/com/novoda/tpbot/human/HumanTelepresenceService.java
@@ -2,7 +2,7 @@ package com.novoda.tpbot.human;
 
 import com.novoda.tpbot.Direction;
 import com.novoda.tpbot.Result;
-import com.novoda.tpbot.support.Observable;
+import com.novoda.support.Observable;
 
 interface HumanTelepresenceService {
