+++ TelepresenceBot/core/src/test/java/com/novoda/support/ObservableTest.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.support;
+package com.novoda.support;
 
 import com.novoda.tpbot.Result;
