+++ TelepresenceBot/core/src/test/java/com/novoda/tpbot/bot/BotPresenterTest.java
@@ -2,7 +2,7 @@ package com.novoda.tpbot.bot;
 
 import com.novoda.tpbot.Direction;
 import com.novoda.tpbot.Result;
-import com.novoda.tpbot.support.Observable;
+import com.novoda.support.Observable;
 
 import org.junit.Before;
 import org.junit.Rule;
