+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/human/HumanActivity.java
@@ -10,9 +10,9 @@ import com.novoda.tpbot.R;
 import com.novoda.tpbot.controls.CommandRepeater;
 import com.novoda.tpbot.controls.ControllerListener;
 import com.novoda.tpbot.controls.ControllerView;
-import com.novoda.tpbot.support.SelfDestructingMessageView;
+import com.novoda.support.SelfDestructingMessageView;
 import com.novoda.tpbot.ServerDeclarationListener;
-import com.novoda.tpbot.support.SwitchableView;
+import com.novoda.support.SwitchableView;
 
 public class HumanActivity extends AppCompatActivity implements HumanView {
