+++ TelepresenceBot/mobile/src/main/res/layout-land/activity_human.xml
@@ -7,7 +7,7 @@
   android:orientation="horizontal"
   tools:context="com.novoda.tpbot.human.HumanActivity">
 
-  <com.novoda.tpbot.support.SelfDestructingMessageView
+  <com.novoda.support.SelfDestructingMessageView
     android:id="@+id/bot_controller_debug_view"
     style="@style/Controller.DebugText"
     android:layout_width="0dp"
@@ -15,7 +15,7 @@
     android:layout_weight="50"
     tools:text="This is some text yo" />
 
-  <com.novoda.tpbot.support.SwitchableView
+  <com.novoda.support.SwitchableView
     android:id="@+id/bot_switchable_view"
     android:layout_width="0dp"
     android:layout_height="match_parent"
