+++ TelepresenceBot/mobile/src/main/res/layout/activity_landing.xml
@@ -7,14 +7,14 @@
   tools:context="com.novoda.tpbot.landing.LandingActivity"
   android:orientation="vertical">
 
-  <com.novoda.tpbot.support.TextViewWithForeground
+  <com.novoda.support.TextViewWithForeground
     android:id="@+id/bot_selection"
     style="@style/Landing.Bot"
     android:layout_width="match_parent"
     android:layout_height="0dp"
     android:layout_weight="100" />
 
-  <com.novoda.tpbot.support.TextViewWithForeground
+  <com.novoda.support.TextViewWithForeground
     android:id="@+id/human_selection"
     style="@style/Landing.Human"
     android:layout_width="match_parent"
